refactor: Moved event contents to top-level, renamed event to CloudEvent
