+++ mock.go
@@ -1,7 +1,10 @@
 package orchestrator
 
 import (
+	"encoding/base64"
 	"encoding/json"
+
+	cloudevent "github.com/cloudevents/sdk-go/v2/event"
 )
 
 const DefaultMockRequestID = "mockid"
@@ -66,4 +69,35 @@ func NewMockRequest(manifest any, opts ...MockRequestOption) (*Request, error) {
 	}
 
 	return req, err
+}
+
+func NewMockCloudEvent(manifest any, opts ...MockRequestOption) (*cloudevent.Event, error) {
+	req, err := NewMockRequest(manifest, opts...)
+	if err != nil {
+		return nil, err
+	}
+
+	reqdata, err := json.Marshal(req)
+	if err != nil {
+		return nil, err
+	}
+
+	buf := make([]byte, base64.StdEncoding.EncodedLen(len(reqdata)))
+	base64.StdEncoding.Encode(buf, reqdata)
+	data, err := json.Marshal(&CloudEventData{
+		Message: PubSubMessage{
+			Data:        reqdata,
+			ID:          "id",
+			PublishTime: "time",
+			Attributes:  PubSubMessageAttributes{},
+		},
+		Subscription: "sub",
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	e := cloudevent.New(cloudevent.CloudEventsVersionV03)
+	e.DataEncoded = data
+	return &e, nil
 }
